requirements.txt
spotipy
youtube-dl
python-dotenv
